http basic, http-app, express-basic were explored
